+++ offchain/crates/solana-client-tools/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod account;
 pub mod instruction;
 pub mod keypair;
-pub mod log;
 pub mod payer;
 pub mod rpc;
 pub mod transaction;