+++ offchain/crates/solana-admin-cli/sol-conversion/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- fix println ([#226](https://github.com/doublezerofoundation/doublezero-offchain/pull/226))
+
 ## [0.0.1](https://github.com/doublezerofoundation/doublezero-offchain/compare/doublezero-sol-conversion-admin-cli/v0.0.0...doublezero-sol-conversion-admin-cli/v0.0.1) - 2025-10-21
 
 ### Other