+++ offchain/crates/solana-cli/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- use tracing for `revenue-distribution relay` commands ([#226](https://github.com/doublezerofoundation/doublezero-offchain/pull/226))
+
+## [0.3.0](https://github.com/doublezerofoundation/doublezero-offchain/releases/tag/doublezero-solana/v0.3.0) - 2025-11-24
+
 - uptick to v0.3.0 ([#210](https://github.com/doublezerofoundation/doublezero-offchain/pull/210))
 - add `revenue-distribution fetch validator-debts` command ([#201](https://github.com/doublezerofoundation/doublezero-offchain/pull/201))
 - solana-cli: add shared validator access validation for `prepare-validator-access` and `request-validator-access` commands, with a hidden `--force` flag on the request command ([#211](https://github.com/doublezerofoundation/doublezero-offchain/pull/211))